* master:
  migrations: Fix for MySQL 5.5 preferring 'engine' over 'type'.
  migrations: Add comments about one-to-many collection names.
  util: Tick lombok.
  docs: Fix build.properties instructions.
  domain: Fix findAllIds to use Longs.
  docs: Working on getting started.
  all: Remove unneeded webdriver repo from ivysettings.
  starter: Add buildfile.
  starter: Set joist.version to 1.0.1.
  starter: Remove unneeded repos from ivysettings.
  docs: Index edits.
  docs: Use 2-space indent, move caveat.
  docs: Fix github links.
  features: Change child ownership, test cacading delete.
  migrations: Add MySQL 'int' type to timeandmoney rules.
  starter: Update EmployeeTest.
  perf: Remove bindgen tests.
  domain: Refactor early failure to dedent.

+++ domain/src/main/java/joist/domain/AbstractQueries.java
@@ -33,7 +33,7 @@ public abstract class AbstractQueries<T extends DomainObject> {
     return Select.from(this.aliasType).count();
   }
 
-  public List<Integer> findAllIds() {
+  public List<Long> findAllIds() {
     return Select.from(this.aliasType).listIds();
   }
